requirements.txt
Pillow
pycryptodome
click